requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
python-decouple==3.8
alembic==1.13.1
websockets==12.0
python-multipart==0.0.6
